feat(auth): support render prop
There are many use cases where we want to render Component even if the autheorization failed. (Such as disabled
action button if the user is not permitted etc..). This commit add the
ability to pass render function to AuthorizedContent component which in
turn pass the isAuthorized value and render the component. this way the
user can decide which component to render depend on isAuthorized value.